# storage/api-docs/user-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PayCan User API Documentation",
        "description": "User API for PayCan payment processing platform. Access products, manage orders, subscriptions, and user profile. Requires JWT authentication via Sanctum.",
        "contact": {
            "name": "PayCan Support",
            "email": "support@paycan.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "PayCan API Server"
        }
    ],
    "paths": {
...
        "/api/user/wallets": {
            "get": {
                "tags": [
                    "User Wallets"
                ],
                "summary": "Get user wallets",
                "description": "Retrieve all credit wallets and current balances for the authenticated user",
                "operationId": "getUserWallets",
                "responses": {
                    "200": {
                        "description": "Wallets retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Wallet"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/wallets/{type}": {
            "get": {
                "tags": [
                    "User Wallets"
                ],
                "summary": "Get specific user wallet",
                "description": "Retrieve details and current balance for a specific wallet type (e.g. basic, premium)",
                "operationId": "getUserWalletByType",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Wallet type (e.g. basic, premium)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "premium"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wallet retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Wallet"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/wallets/{type}/transactions": {
            "get": {
                "tags": [
                    "User Wallets"
                ],
                "summary": "Get user wallet transactions",
                "description": "Retrieve paginated transaction history for the authenticated user's specific wallet",
                "operationId": "getUserWalletTransactions",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Wallet type (e.g. basic, premium)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "filter[type]",
                        "in": "query",
                        "description": "Filter direction (credit, debit)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "credit",
                                "debit"
                            ]
                        }
                    },
                    {
                        "name": "filter[action]",
                        "in": "query",
                        "description": "Filter action",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wallet transactions retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/WalletTransaction"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/wallets/transactions": {
            "get": {
                "tags": [
                    "User Wallets"
                ],
                "summary": "Get all wallet transactions for authenticated user",
                "description": "Retrieve paginated transaction history across all credit wallets for the authenticated user",
                "operationId": "getUserAllWalletTransactions",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "filter[wallet_type]",
                        "in": "query",
                        "description": "Filter by wallet type (basic, premium)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[type]",
                        "in": "query",
                        "description": "Filter direction (credit, debit)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "credit",
                                "debit"
                            ]
                        }
                    },
                    {
                        "name": "filter[action]",
                        "in": "query",
                        "description": "Filter action",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[reference_id]",
                        "in": "query",
                        "description": "Filter reference ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transactions retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/WalletTransaction"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/wallets/usages": {
            "get": {
                "tags": [
                    "User Wallets"
                ],
                "summary": "Get credit usage logs for authenticated user",
                "description": "Retrieve paginated credit usage logs (e.g. AI agent tasks, prompts) for authenticated user",
                "operationId": "getUserWalletUsages",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "filter[wallet_type]",
                        "in": "query",
                        "description": "Filter by wallet type (basic, premium)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[reference_id]",
                        "in": "query",
                        "description": "Filter reference ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usage logs retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/WalletTransaction"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Wallet": {
                "title": "Wallet",
                "description": "User credit wallet model",
                "required": [
                    "id",
                    "user_id",
                    "type",
                    "balance",
                    "currency",
                    "is_active"
                ],
                "properties": {
                    "id": {
                        "description": "Wallet ID (ULID)",
                        "type": "string",
                        "example": "01JM7A9B8C7D6E5F4G3H2J1K0L"
                    },
                    "user_id": {
                        "description": "User ID",
                        "type": "string",
                        "example": "01HKXZ7K5QGXP0B1J2R3T4V5W6"
                    },
                    "type": {
                        "description": "Wallet type (e.g. basic, premium, or custom)",
                        "type": "string",
                        "enum": [
                            "basic",
                            "premium"
                        ],
                        "example": "basic"
                    },
                    "balance": {
                        "description": "Current credit balance",
                        "type": "number",
                        "format": "float",
                        "example": 1250
                    },
                    "currency": {
                        "description": "Unit of account",
                        "type": "string",
                        "example": "credits"
                    },
                    "is_active": {
                        "description": "Whether wallet is active",
                        "type": "boolean",
                        "example": true
                    },
                    "meta": {
                        "description": "Additional wallet metadata",
                        "type": "object",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "USER AUTHENTICATION (Sanctum JWT) - For mobile apps & SPAs. To get a user token for testing, use the Admin API endpoint: POST /api/admin/users/sync (see Admin API documentation). Format: Bearer <user_jwt_token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}